+++ exporters/autoexport/registry_test.go
@@ -17,12 +17,12 @@ package autoexport
 import (
 	"context"
 	"fmt"
+	"sync"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
-	"go.opentelemetry.io/otel/exporters/otlp/otlptrace"
 	"go.opentelemetry.io/otel/exporters/stdout/stdouttrace"
 	"go.opentelemetry.io/otel/sdk/trace"
 )
@@ -59,41 +59,50 @@ func TestRegistryIsConcurrentSafe(t *testing.T) {
 		require.NoError(t, r.store(exporterName, stdoutFactory))
 	})
 
+	var wg sync.WaitGroup
+
+	wg.Add(1)
 	go func() {
+		defer wg.Done()
 		assert.NotPanics(t, func() {
 			require.ErrorIs(t, r.store(exporterName, stdoutFactory), errDuplicateRegistration)
 		})
 	}()
 
+	wg.Add(1)
 	go func() {
+		defer wg.Done()
 		assert.NotPanics(t, func() {
 			exp, err := r.load(context.Background(), exporterName)
-			assert.Nil(t, err, "missing exporter in registry")
+			assert.NoError(t, err, "missing exporter in registry")
 			assert.IsType(t, &stdouttrace.Exporter{}, exp)
 		})
 	}()
+
+	wg.Wait()
 }
 
 func TestSubsequentCallsToGetExporterReturnsNewInstances(t *testing.T) {
 	const exporterType = "otlp"
 	exp1, err := spanExporter(context.Background(), exporterType)
-	assert.Nil(t, err)
-	assert.IsType(t, &otlptrace.Exporter{}, exp1)
+	assert.NoError(t, err)
+	assertOTLPHTTPExporter(t, exp1)
 
 	exp2, err := spanExporter(context.Background(), exporterType)
-	assert.Nil(t, err)
-	assert.IsType(t, &otlptrace.Exporter{}, exp2)
+	assert.NoError(t, err)
+	assertOTLPHTTPExporter(t, exp2)
+
 	assert.NotSame(t, exp1, exp2)
 }
 
 func TestDefaultOTLPExporterFactoriesAreAutomaticallyRegistered(t *testing.T) {
 	exp1, err := spanExporter(context.Background(), "")
 	assert.Nil(t, err)
-	assert.IsType(t, &otlptrace.Exporter{}, exp1)
+	assertOTLPHTTPExporter(t, exp1)
 
 	exp2, err := spanExporter(context.Background(), "otlp")
 	assert.Nil(t, err)
-	assert.IsType(t, &otlptrace.Exporter{}, exp2)
+	assertOTLPHTTPExporter(t, exp2)
 }
 
 func TestEnvRegistryCanRegisterExporterFactory(t *testing.T) {